Refactoring out deployment
